refactor: event finding in roll
